+++ composition/main.tf
@@ -4,5 +4,6 @@ module "infrastructure" {
     for node in var.nodes : node.domain_name => node
   }
 
-  node = each.value
+  node           = each.value
+  ssh_public_key = file(pathexpand(var.ssh_public_key_path))
 }