+++ composition/variables.tf
@@ -18,5 +18,14 @@ variable "nodes" {
     cpu                = any
     features           = any
     devices            = any
+    instance_id        = string
+    hostname           = string
+    cloudinit_name     = string
   }))
+}
+
+variable "ssh_public_key_path" {
+  description = "Path to the SSH public key used by cloud-init"
+  type        = string
+  default     = "~/.ssh/id_ed25519.pub"
 }