+++ resource_module/libvirt_cloudinit_disk/outputs.tf
@@ -1,0 +1,9 @@
+output "cloudinit_name" {
+  description = "Name of the created libvirt cloudinit disk"
+  value       = libvirt_cloudinit_disk.init.name
+}
+
+output "cloudinit_path" {
+  description = "Path of the created libvirt cloudinit disk"
+  value       = libvirt_cloudinit_disk.init.path
+}