+++ resource_module/libvirt_cloudinit_volume/outputs.tf
@@ -1,0 +1,4 @@
+output "volume_name" {
+  description = "Name of the cloud-init volume"
+  value       = libvirt_volume.cloudinit.name
+}